+++ src/environments/types/config.ts
@@ -232,6 +232,7 @@ export const createEnvironmentConfig = <
   transport: Transport;
   indexerUrl: string;
   governanceIndexerUrl: string;
+  lunarIndexerUrl?: string;
   tokens: TokensConfig<tokens>;
   markets: MarketsConfig<markets, tokens>;
   vaults: VaultsConfig<vaults, tokens>;
@@ -495,6 +496,7 @@ export const createEnvironmentConfig = <
     chain: config.chain,
     indexerUrl: config.indexerUrl,
     governanceIndexerUrl: config.governanceIndexerUrl,
+    lunarIndexerUrl: config.lunarIndexerUrl,
     tokens: tokenContracts,
     markets: marketContracts,
     vaults: vaultsContracts,
@@ -532,6 +534,7 @@ export type Environment<
   chain: Chain;
   indexerUrl: string;
   governanceIndexerUrl: string;
+  lunarIndexerUrl?: string;
   tokens: {
     [name in keyof tokens]: TokenContractReturnType;
   };